Use 'dp' from precision_module everywhere
One should not use "kind=8" in modern Fortran, but rather have it as a
parameter declared at one place.

+++ src/1d/hyperbolic_step.f90
@@ -17,7 +17,7 @@ subroutine hyperbolic_step(solution,solver)
     ! Local storage
     integer :: i, m, mw
     type(geometry_type) :: geometry
-    real(kind=8) :: cfl
+    real(dp) :: cfl
 
     ! Take apart solution and solver data types for ease of reference below
     associate(q => solution%q, aux => solution%aux, f => solver%f,           &
